+++ certbot-dns-dnsmadeeasy/certbot_dns_dnsmadeeasy/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_dnsmadeeasy.dns_dnsmadeeasy` plugin.
